requirements.txt
flask
Flask-PyMongo
flask-restful
voluptuous==0.10.5
